Don't pass domain into phase1 network configuration
The only reason we are passing the domain object into network configuration is
to allow it to mutate the domain object before passed to libvirt. This is of no
need when phase1 is being executed in virt-handler that does not define the
domain for libvirt. Instead of trying to make virt-handler to regenerate the
domain to satisfy network configuration api, rework phase1 code to not require
the domain in the first place.

This patch does not remove the domain argument from phase1 yet but instead
passes nil. The next patch will actually remove the no longer used argument.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>