Add build, deployment and test support for ppc64le arch
This commit adds support in Kubevirt for building, deploying and testing
VMs for IBM Power Systems little-endian architecture (ppc64le).

A .bazelrc file was created in order to move common and architecture
dependent options to it, this makes it easier to add new architectures
support in the future, the architecture dependent options are selected at
runtime by checking the output of uname -m and passing it to the --config
option added. The .bazelrc file generated at runtime by the CI was renamed
to ci.bazelrc and made and untracked file, the .bazelrc added attempts to
import the ci.bazelrc in case it exists.

The WORKSPACE file now contains pointers to the ppc64le specific
dependencies in order to properly build kubevirt and a select was added to
the BUILD.bazel files that use them to use the appropriate ones depending
on the architecture in which Kubevirt is build, the default dependencies
for an unmatched architecture is to use the amd64 ones that were already
defined by Kubevirt.

As part of the WORKSPACE a ppc64le binary puller was added given that
after the migration of rules_docker to a Go implementation the support for
container_pull was broken for architectures that are not amd64, due to the
binary the tool uses. So whenever pulling a container in ppc64le the
rebuilt puller is used by pointing it in the puller_linux attribute.

The .travis.yml file was modified to add ppc64le arch as well, but for the
deploy steps of it they are currently only done in amd64. Some of the tests
had their expected output adjusted to support this new arch. Two test cases
to note are the ones that check privileged mode, for now in Power we are
running in it so these two test cases are currently being skipped.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>