Add autoattachSerialConsole flag
For completeness also allow opting out of the default serial console
device, like we allow for VNC.

For some usecases on some Operating Systems an unneeded device can have
unwanted side-effects. The most recent case we saw is Windows, where
every user in the Guest OS by default sees the serial console as an
unpluggable PCI device.

First this is confusing for the user and second the serial console does
not even serve any purpose on windows guests.

Signed-off-by: Roman Mohr <rmohr@redhat.com>