Add annotation to place pci addresses on root bus
If a VMI is started with the annotation
"kubevirt.io/placePCIDevicesOnRootComplex": "true", all PCI devices are
attached to the root PCI bus. This makes them not hot-unplugable in
windows.

Further the functional tests check if for the most common devices, no
additional pci bus is added. When the devices are present and only the
root bus exists, they are by definition on the root bus.

Signed-off-by: Roman Mohr <rmohr@redhat.com>