Set kubevirt priority class on all core components
As of K8s 1.16, the `scheduler.alpha.kubernetes.io/critical-pod`
annotation is deprecated. As a result, we now deploy a priority class
for our core components. Since the operator is managing the cert
rotation for all of our components, we also consider it as a high
priority component and in result, the priority class must be deployed
before the operator is being deployed.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>