Fix tests relying on bearer token behavior
A handful of functional tests relied on using "--token" to indicate preference
as to which service account the test should run as. These tests might have
worked at some point in the past, but don't behave correctly with current k8s
or openshift clients.

Since the intention of the test is to simply verify VM creation
behavior/permissions, these tests have been changed to instead use the
"auth can-i" subcommand--which unambiguously checks permissions for the
explicitly named account.

Signed-off-by: Stu Gott <sgott@redhat.com>