Don't use a builtin taint
Some tests in our test suite use taints to trigger a migration. This is
normally fine. However, in deployments where the node maintenance
operator is installed, the node.kubernetes.io/unschedulable built-in
taint is used. This taint cannot be applied manually. If this is the
case, we can simply use another key.

Signed-off-by: Stu Gott <sgott@redhat.com>