Create scripts for building and testing from prs
We add separate Dockerfile, test and build scripts for pr
to avoid interference with changes in nightly build.

Within the test script an ephemeral registry is created on the test
cluster. The machine config is patched to allow this insecure registry.

Then the images are built. A port forwarding is established to be able
to push against the ephemeral registry, and the images are pushed against
localhost.

Then the manifests are created using the exported route for the local
ephemeral registry. Kubevirt is deployed using the created manifests.

At last the tests that have been compiled during the image build are
started.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>