Handle DataVolume volumes mistakenly used as PVC volumes
When a PVC created by a DataVolume is used as volume source, it can
cause problems with VM startup (as the Data Volume may not be ready
yet). This commit adds a check to detect the erronous situation where a
PVC created by a Data Volume is used directly as Volume Source.

Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>