Preparing kubevirt to make the ipv6 lane mandatory
Since ipv6 support was merged to kubevirt there were servel PRs that broke
the ipv6 lane.
Since the lane is not mandaroty it is very hard to track and to catch
those failures on time.

To avoid this situation in the future. The current failing tests on the
ipv6 lane are skipped so the lane can become green.
Once this commit will be merged, the ipv6 lane can be moved to be
mandatory.

Signed-off-by: Alona Kaplan <alkaplan@redhat.com>