travis-ci: Extract the script to an external file
Cleanup the travis.yml file by extracting the script to an external
file.

It should be now safe to use the `-e` options to fail early on any error
during the run.
In addition, it may be useful to run the same script locally to debug
issues detected on CI.

Signed-off-by: Edward Haas <edwardh@redhat.com>