Revert "Merge pull request #3314 from maya-r/loopdev"
This reverts commit 5c0e98b816c228ba312cd6f3350f2a23a7ae8d4a, reversing
changes made to 7b7e02d8bf393549a21afcc0c871d52099afd20c.

The `losetup` is not namespaced, and as such is creating a new loop
device in the host for every time the `disks-images-provider` container
is created. As a result, our CI is doomed to only work 100 times on each
host (the number of pre-provisioned devices). Once the 'next available'
loop device is not a pre-provisioned one, the `disks-images-provider`
container will fail to reach a `Running` state, which will cause the
functest make target to fail after 5 minutes (default timeout of the
test script).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>