Fix flaky vmi watch test
`kubevirt get vmi -w` can take a while to connect. If it takes too long, some events will be missed.
Since there no way to know when it's connected (even with a big -v there's nothing in stderr).
Added a "guard vmi", which triggers initial output that we wait for, ensuring kubectl is connected.

There was also an issue with the VMI not necessarily having an IP at the "scheduled" stage.
Removed the check for it.

Signed-off-by: Jed Lejosne <jed@redhat.com>