Add verification targets to the Makefile
The purpose of those targets is to serve as an entrypoint for Prow jobs.
The idea is to eventually replace automation/travisci-test.sh and the
single Travis-CI job with multiple Prow jobs that will run in parallel.

I also modified the travis-ci testing script in this commit to get proper CI
coverage for this work. There should be no change in the functionality
of the targets as opposed to calling them and wrapping the calls with
some checks from a shell script.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>